requirements-api.txt
# Slim requirements for the FastAPI API server (Render deployment).
# Excludes spaCy, OpenAI, and ddgs which are only needed for pipeline scripts.

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # Postgres driver (Neon)

# Auth (Phase 6F)
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Config
python-dotenv>=1.0.0
